member5_dashboard_oversight/requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.0.0
pytest>=8.0.0